requirements.txt
transformers[torch]
pillow
IPython
datasets
sentence-transformers[train]